requirements.txt
langchain
openai
streamlit
python-dotenv
langchain_community

